Friendship controller 수정

+++ src/api/friendship/controllers/friendship.ts
@@ -27,21 +27,25 @@ export default factories.createCoreController(
           {
             filters: {
               $or: [
-                { friend_confirm: userId },
-                { friend_request: userId },
-                { block: userId },
-                { blocked_by: userId },
-                { friend_confirm: +friendId },
-                { friend_request: +friendId },
-                { block: +friendId },
-                { blocked_by: +friendId },
+                {
+                  $and: [
+                    { follow_sender: userId },
+                    { follow_receiver: +friendId },
+                  ],
+                },
+                {
+                  $and: [
+                    { follow_receiver: userId },
+                    { follow_sender: +friendId },
+                  ],
+                },
               ],
             },
             populate: {
-              friend_confirm: { fields: "id" },
-              friend_request: { fields: "id" },
-              block: { fields: "id" },
-              blocked_by: { fields: "id" },
+              follow_receiver: { fields: ["nickname"] },
+              follow_sender: { fields: ["nickname"] },
+              block: { fields: ["nickname"] },
+              blocked: { fields: ["nickname"] },
             },
           }
         );
@@ -49,19 +53,22 @@ export default factories.createCoreController(
         let modifiedFriendship;
 
         if (
-          friendship[0].status === "pending" ||
-          friendship[0].status === "friend"
+          friendship[0].status === "PENDING" ||
+          friendship[0].status === "FRIEND"
         ) {
           modifiedFriendship = {
             status: friendship[0].status,
-            friend_confirm: friendship[0].friend_confirm,
-            friend_request: friendship[0].friend_request,
+            follow_receiver: friendship[0].follow_receiver,
+            follow_sender: friendship[0].follow_sender,
           };
-        } else if (friendship[0].status === "block") {
+        } else if (
+          friendship[0].status === "BLOCK_ONE" ||
+          friendship[0].status === "BLOCK_BOTH"
+        ) {
           modifiedFriendship = {
             status: friendship[0].status,
             block: friendship[0].block,
-            blocked_by: friendship[0].blocked_by,
+            blocked: friendship[0].blocked,
           };
         }
 
@@ -72,6 +79,7 @@ export default factories.createCoreController(
     },
 
     // 관계 만들기 (친구 요청)
+    // 친분 있으면 못 만들게
     async create(ctx) {
       if (!ctx.state.user) {
         return ctx.UnauthorizedError("권한이 없습니다.");
@@ -80,18 +88,83 @@ export default factories.createCoreController(
         return ctx.badRequest("friendId가 필요합니다.");
       }
       try {
-        const newFriendship = await strapi.entityService.create(
+        const { id: userId, nickname: userNickname } = ctx.state.user;
+        const { friendId } = ctx.query;
+
+        // 친분있는지 유무
+        const friendship = await strapi.entityService.findMany(
           "api::friendship.friendship",
           {
-            data: {
-              status: "pending",
-              friend_confirm: ctx.state.user.id,
-              friend_request: ctx.query.friendId,
+            filters: {
+              $or: [
+                {
+                  $and: [
+                    { follow_sender: userId },
+                    { follow_receiver: +friendId },
+                  ],
+                },
+                {
+                  $and: [
+                    { follow_receiver: userId },
+                    { follow_sender: +friendId },
+                  ],
+                },
+              ],
+            },
+            populate: {
+              follow_receiver: { fields: ["nickname"] },
+              follow_sender: { fields: ["nickname"] },
+              block: { fields: ["nickname"] },
+              blocked: { fields: ["nickname"] },
             },
           }
         );
 
-        ctx.send("친구 요청을 보냈습니다.");
+        if (friendship.length !== 0) {
+          return ctx.badRequest("이미 follow 하고 있습니다.");
+        }
+
+        const newFriendship = await strapi.entityService.create(
+          "api::friendship.friendship",
+          {
+            data: {
+              status: "PENDING",
+              follow_receiver: +friendId,
+              follow_sender: userId,
+            },
+          }
+        );
+
+        // 친구 신청 알림
+        if (newFriendship) {
+          const notice = await strapi.entityService.findMany(
+            "api::notice.notice",
+            {
+              filters: {
+                sender: userId,
+                receiver: +friendId,
+              },
+            }
+          );
+
+          if (notice.length !== 0) {
+            return ctx.badRequest("이미 follow 요청을 보냈습니다.");
+          }
+
+          const newNotice = await strapi.entityService.create(
+            "api::notice.notice",
+            {
+              data: {
+                body: `${userNickname}님이 친구 요청을 보냈습니다.`,
+                receiver: +friendId,
+                sender: userId,
+                event: "FRIEND",
+              },
+            }
+          );
+
+          ctx.send("친구 요청을 보냈습니다.");
+        }
       } catch (e) {
         console.log(e);
       }
@@ -112,73 +185,126 @@ export default factories.createCoreController(
         return ctx.badRequest("friendId, status가 필요합니다.");
       }
       try {
+        // 친분 유무
         const friendship = await strapi.entityService.findMany(
           "api::friendship.friendship",
           {
             filters: {
               $or: [
-                { friend_confirm: userId },
-                { friend_request: userId },
-                { block: userId },
-                { blocked_by: userId },
-                { friend_confirm: +friendId },
-                { friend_request: +friendId },
-                { block: +friendId },
-                { blocked_by: +friendId },
+                {
+                  $and: [
+                    { follow_sender: userId },
+                    { follow_receiver: +friendId },
+                  ],
+                },
+                {
+                  $and: [
+                    { follow_receiver: userId },
+                    { follow_sender: +friendId },
+                  ],
+                },
               ],
             },
             populate: {
-              friend_confirm: { fields: "id" },
-              friend_request: { fields: "id" },
-              block: { fields: "id" },
-              blocked_by: { fields: "id" },
+              follow_receiver: { fields: ["nickname"] },
+              follow_sender: { fields: ["nickname"] },
+              block: { fields: ["nickname"] },
+              blocked: { fields: ["nickname"] },
             },
           }
         );
 
-        // pending -> friend
-        if (friendship[0].status === "pending" && status === "friend") {
-          const updatedFriendship = await strapi.entityService.update(
-            "api::friendship.friendship",
-            friendship[0].id
-            // 상상코딩
-            // { data: { status: "friend" } }
-          );
-          // friend -> block
-        } else if (friendship[0].status === "friend" && status === "block") {
-          const updatedFriendship = await strapi.entityService.update(
-            "api::friendship.friendship",
-            friendship[0]
-            // 상상코딩
-            // {
-            //   data: {
+        if (friendship.length === 0) {
+          return ctx.badRequest("친분이 없는 상태입니다.");
+        }
 
-            // status: "block",
-            // block: { connect: [{ id: userId }] },
-            // blocked_by: { connect: [{ id: +friendId }] },
-            //   },
-            // }
-          );
+        // // pending -> friend
+        // // 상대방이 팔로우 요청을 보냈다면 수락 하는 단계 (맞팔)
+        if (
+          friendship[0].status === "PENDING" &&
+          status === "friend" &&
+          friendship[0].follow_receiver.id === userId &&
+          friendship[0].follow_sender.id === +friendId
+        ) {
+          try {
+            const updatedFriendship = await strapi.entityService.update(
+              "api::friendship.friendship",
+              friendship[0].id,
+              { data: { ...friendship[0], status: "FRIEND" } }
+            );
+          } catch (e) {
+            ctx.badRequest(
+              "Failed to update friendship. Invalid date or missing required fields."
+            );
+          }
+
+          // friend -> block
+          // 친구 중 상태에서 한명이 block할때
+        } else if (
+          friendship[0].status === "FRIEND" ||
+          (friendship[0].status === "BLOCK_ONE" && status === "block")
+        ) {
+          try {
+            const updatedFriendship = await strapi.entityService.update(
+              "api::friendship.friendship",
+              friendship[0].id,
+              {
+                data: {
+                  ...friendship[0],
+                  status:
+                    friendship[0].block.length === 0
+                      ? "BLOCK_ONE"
+                      : "BLOCK_BOTH",
+                  block: { connect: [userId] },
+                  blocked: { connect: [+friendId] },
+                },
+              }
+            );
+          } catch (e) {
+            ctx.badRequest(
+              "Failed to update friendship. Invalid date or missing required fields."
+            );
+          }
 
           // block -> friend
-        } else if (friendship[0].status === "block" && status === "friend") {
-          const updatedFriendship = await strapi.entityService.update(
-            "api::friendship.friendship",
-            friendship[0].id
-            // 상상 코딩
-            // {
-            //   data: {
-            //     status: "friend",
-            //     block: { disconnect: [{ id: userId }, { id: +friendId }] },
-            //     blocked_by: { disconnect: [{ id: userId }, { id: +friendId }] },
-            //   },
-            // }
-          );
+          // block 해제
+        } else if (
+          friendship[0].status === "BLOCK_ONE" ||
+          (friendship[0].status === "BLOCK_BOTH" && status === "friend")
+        ) {
+          try {
+            const updatedFriendship = await strapi.entityService.update(
+              "api::friendship.friendship",
+              friendship[0].id,
+
+              {
+                data: {
+                  ...friendship[0],
+                  status:
+                    friendship[0].status === "BLOCK_BOTH"
+                      ? "BLOCK_ONE"
+                      : "FRIEND",
+                  block: {
+                    disconnect: [userId],
+                  },
+                  blocked: {
+                    disconnect: [+friendId],
+                  },
+                },
+              }
+            );
+          } catch (e) {
+            ctx.badRequest(
+              "Failed to update friendship. Invalid date or missing required fields."
+            );
+          }
+        } else {
+          return ctx.badRequest("Failed to update friendship");
         }
 
         ctx.send("Update Friendship Success");
       } catch (e) {
-        console.log(e);
+        return ctx.badRequest("Failed to update friendship");
       }
     },
 
@@ -191,25 +317,38 @@ export default factories.createCoreController(
           {
             filters: {
               $or: [
-                { friend_confirm: userId },
-                { friend_request: userId },
-                { block: userId },
-                { blocked_by: userId },
-                { friend_confirm: +friendId },
-                { friend_request: +friendId },
-                { block: +friendId },
-                { blocked_by: +friendId },
+                {
+                  $and: [
+                    { follow_sender: userId },
+                    { follow_receiver: +friendId },
+                  ],
+                },
+                {
+                  $and: [
+                    { follow_receiver: userId },
+                    { follow_sender: +friendId },
+                  ],
+                },
               ],
+            },
+            populate: {
+              follow_receiver: { fields: ["nickname"] },
+              follow_sender: { fields: ["nickname"] },
+              block_receiver: { fields: ["nickname"] },
+              block_sender: { fields: ["nickname"] },
             },
           }
         );
 
-        const deleteFriendship = await strapi.entityService.delete(
-          "api::friendship.friendship",
-          friendship[0].id
-        );
-
-        ctx.send(`${userId}와 ${+friendId}의 친구 관계가 삭제 되었습니다.`);
+        if (friendship.length !== 0) {
+          const deleteFriendship = await strapi.entityService.delete(
+            "api::friendship.friendship",
+            friendship[0].id
+          );
+          return ctx.send(
+            `${userId}와 ${+friendId}의 친구 관계가 삭제 되었습니다.`
+          );
+        }
       } catch (e) {
         console.log(e);
       }
